extract common hover logic to another file
